fix missed integer comparison warning

+++ duf.hpp
@@ -52,7 +52,7 @@ auto group_by(const Container& df, Members... members)
     // at each insertion).
     auto groub_elements_counter = std::map<key_type, std::vector<int>>();
 
-    for(int i = 0; i < df.size(); i++)
+    for(size_t i = 0; i < df.size(); i++)
     {
         const auto key = std::make_tuple((df[i].*members)...);
         groub_elements_counter[key].emplace_back(i);
